refactor(viewer): use type import
