+++ libs/annotation/src/components/AnnotationOverlay/utils/getIsIntersection.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../types'
+import type { Point } from '../../../types'
 import { getIsIntersection } from './getIsIntersection'
 
 describe('getIsIntersection:', () => {
