+++ libs/annotation/src/components/AnnotationOverlay/utils/getIsIntersection.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../types'
+import type { Point } from '../../../types'
 
 /**
  * Determine if ab and cd intersect
