+++ libs/annotation/src/components/TextDrawer/Typing.types.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../types'
+import type { Point } from '../../types'
 
 export interface TypingProps {
   points: Point[]
