+++ libs/annotation/src/hooks/useCreatingDrawableAnnotation/utils/getArrowPosition.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../../types'
+import type { Point } from '../../../types'
 import { getArrowPosition } from './getArrowPosition'
 
 describe('getArrowPosition:', () => {
