+++ libs/annotation/src/hooks/useRulerMeasurement/utils/getRulerConnectingLine.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../../types'
+import type { Point } from '../../../types'
 import { getRulerConnectingLine } from './getRulerConnectingLine'
 
 describe('getConnectingLinePoints: ', () => {
