+++ libs/annotation/src/mocks/utils.ts
@@ -1,3 +1,3 @@
-import { Point } from '../types'
+import type { Point } from '../types'
 
 export const mockPixelToCanvas = ([x, y]: Point): Point => [x, y]
