+++ libs/annotation/src/utils/getEditingTextPosition.ts
@@ -1,4 +1,4 @@
-import { Point } from '../types'
+import type { Point } from '../types'
 
 export function getEditingTextPosition(currentPoint: Point): Point {
   return currentPoint
