+++ libs/annotation/src/utils/modifyConnectingLine.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../types'
+import type { Point } from '../types'
 import { modifyConnectingLine } from './modifyConnectingLine'
 
 describe('modifyConnectingLine', () => {
