+++ libs/annotation/src/utils/modifyConnectingLine.ts
@@ -1,4 +1,4 @@
-import { Point } from '../types'
+import type { Point } from '../types'
 
 const getDistance = (pointA: Point, pointB: Point) => {
   return Math.pow(pointA[0] - pointB[0], 2) + Math.pow(pointA[1] - pointB[1], 2)
