+++ libs/annotation/src/utils/stringifyPoints.ts
@@ -1,4 +1,4 @@
-import { Point } from '../types'
+import type { Point } from '../types'
 
 export const stringifyPoints = (points: Point[]): string => {
   return points.map((point) => `${point[0]},${point[1]}`).join(' ')
